+++ CHANGELOG.md
@@ -9,7 +9,7 @@
   - has now a responsive design to support smaller viewports
 - updata data to 3.10.1c (#555)
 - update run iohook only if required (stash navigation)
-- fix quality higher than 20% showing a lower value as max
+- fix quality higher than 20% showing a lower value as max (#584)
 - fix `Oni-Goroshi Charan's Sword` unable to parse (#575)
 
 ## 0.6.17 (2020-04-02)